requirements.txt
langchain>=0.1.0
langgraph>=0.1.0
streamlit>=1.28.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
plotly>=5.0.0
pandas>=2.0.0
langchain-openai>=0.1.0
langchain-community>=0.1.0